+++ server/controllers/dishes.js
@@ -5,6 +5,17 @@ const createDish = async (req, res) => {
 	try {
 		const { dishName, dishIngredients, dishCategory, dishPrice, dishActive } =
 			req.body
+
+		if (
+			!dishName ||
+			!dishIngredients ||
+			!dishCategory ||
+			!dishPrice ||
+			!dishActive
+		) {
+			return res.status(400).send("All fields are required")
+		}
+
 		const response = await pool.query(
 			"INSERT INTO dishes (dish_name, dish_ingredients, dish_category, dish_price, dish_active, product_taxe) VALUES ($1, $2, $3, $4, $5, 5.5)",
 			[dishName, dishIngredients, dishCategory, dishPrice, dishActive]
@@ -12,18 +23,18 @@ const createDish = async (req, res) => {
 		res.status(200).send(response.rows)
 	} catch (err) {
 		console.log(err)
-		res.status(400).send(err)
+		res.status(500).send("Internal server error")
 	}
 }
 
 // get all dishes
-const getDishes = async (req, res) => {
+const getDishes = async (_req, res) => {
 	try {
 		const response = await pool.query("SELECT * FROM dishes")
 		res.status(200).send(response.rows)
 	} catch (err) {
 		console.log(err)
-		res.status(400).send(err)
+		res.status(500).send("Internal server error")
 	}
 }
 
@@ -31,6 +42,11 @@ const getDishes = async (req, res) => {
 const getDish = async (req, res) => {
 	try {
 		const { id } = req.params
+
+		if (!id) {
+			return res.status(400).send("Dish ID is required")
+		}
+
 		const response = await pool.query(
 			"SELECT * FROM dishes WHERE dish_id = $1",
 			[id]
@@ -38,7 +54,7 @@ const getDish = async (req, res) => {
 		res.status(200).send(response.rows[0])
 	} catch (err) {
 		console.log(err)
-		res.status(400).send(err)
+		res.status(500).send("Internal server error")
 	}
 }
 
@@ -46,28 +62,28 @@ const getDish = async (req, res) => {
 const updateDish = async (req, res) => {
 	try {
 		const { id } = req.params
-		const {
-			dishName,
-			dishIngredients,
-			dishCategory,
-			dishPrice,
-			dishActive,
-		} = req.body
+		const { dishName, dishIngredients, dishCategory, dishPrice, dishActive } =
+			req.body
+
+		if (
+			!id ||
+			!dishName ||
+			!dishIngredients ||
+			!dishCategory ||
+			!dishPrice ||
+			!dishActive
+		) {
+			return res.status(400).send("All fields are required")
+		}
+
 		const response = await pool.query(
 			"UPDATE dishes SET dish_name = $1, dish_ingredients = $2, dish_category = $3, dish_price = $4, dish_active = $5, product_taxe = 5.5 WHERE dish_id = $6",
-			[
-				dishName,
-				dishIngredients,
-				dishCategory,
-				dishPrice,
-				dishActive,
-				id,
-			]
+			[dishName, dishIngredients, dishCategory, dishPrice, dishActive, id]
 		)
 		res.status(200).send(response.rows)
 	} catch (err) {
 		console.log(err)
-		res.status(400).send(err)
+		res.status(500).send("Internal server error")
 	}
 }
 
@@ -75,12 +91,17 @@ const deleteDish = async (req, res) => {
 	try {
 		const { id } = req.params
 
+		if (!id) {
+			return res.status(400).send("Dish ID is required")
+		}
+
 		const response = await pool.query("DELETE FROM dishes WHERE dish_id = $1", [
 			id,
 		])
 		res.status(200).send(response.rows)
 	} catch (err) {
 		console.log(err)
+		res.status(500).send("Internal server error")
 	}
 }
