+++ server/controllers/tables.js
@@ -4,6 +4,11 @@ const pool = require("../db")
 const createTable = async (req, res) => {
 	try {
 		const { year, month, day, status, number } = req.body
+
+		if (!year || !month || !day || !status || !number) {
+			return res.status(400).send("All fields are required")
+		}
+
 		const response = await pool.query(
 			"INSERT INTO tables (table_year, table_month, table_day, table_status, table_number) VALUES ($1, $2, $3, $4, $5)",
 			[year, month, day, status, number]
@@ -11,18 +16,18 @@ const createTable = async (req, res) => {
 		res.status(200).send(response.rows)
 	} catch (err) {
 		console.log(err)
-		res.status(400).send(err)
+		res.status(500).send(err)
 	}
 }
 
 // get all tables
-const getTables = async (req, res) => {
+const getTables = async (_req, res) => {
 	try {
 		const response = await pool.query("SELECT * FROM tables")
 		res.status(200).send(response.rows)
 	} catch (err) {
 		console.log(err)
-		res.status(400).send(err)
+		res.status(500).send(err)
 	}
 }
 
@@ -30,6 +35,11 @@ const getTables = async (req, res) => {
 const getTable = async (req, res) => {
 	try {
 		const { id } = req.params
+
+		if (!id) {
+			return res.status(400).send("All fields are required")
+		}
+
 		const response = await pool.query(
 			"SELECT * FROM tables WHERE table_id = $1",
 			[id]
@@ -37,12 +47,12 @@ const getTable = async (req, res) => {
 		res.status(200).send(response.rows[0])
 	} catch (err) {
 		console.log(err)
-		res.status(400).send(err)
+		res.status(500).send(err)
 	}
 }
 
 // get active tables
-const getActiveTables = async (req, res) => {
+const getActiveTables = async (_req, res) => {
 	try {
 		const response = await pool.query(
 			"SELECT * FROM tables WHERE table_status = 'active'"
@@ -50,7 +60,7 @@ const getActiveTables = async (req, res) => {
 		res.status(200).send(response.rows)
 	} catch (err) {
 		console.log(err)
-		res.status(400).send(err)
+		res.status(500).send(err)
 	}
 }
 
@@ -58,8 +68,18 @@ const getActiveTables = async (req, res) => {
 const updateTable = async (req, res) => {
 	try {
 		const { id } = req.params
+
+		if (!id) {
+			return res.status(400).send("All fields are required")
+		}
+
 		const { table_year, table_month, table_day, table_status, table_products } =
 			req.body
+
+		if (!table_year || !table_month || !table_day || !table_status || !table_products) {
+			return res.status(400).send("All fields are required")
+		}
+
 		const response = await pool.query(
 			"UPDATE tables SET table_year = $1, table_month = $2, table_day = $3, table_status = $4, table_products = $5 WHERE table_id = $6",
 			[table_year, table_month, table_day, table_status, table_products, id]
@@ -67,7 +87,7 @@ const updateTable = async (req, res) => {
 		res.status(200).send(response.rows)
 	} catch (err) {
 		console.log(err)
-		res.status(400).send(err)
+		res.status(500).send(err)
 	}
 }
 
@@ -76,6 +96,10 @@ const updateTableStatus = async (req, res) => {
 	try {
 		const { id } = req.params
 		const { table_status, sale_id } = req.body
+
+		if (!id || !table_status || !sale_id) {
+			return res.status(400).send("All fields are required")
+		}
 
 		let response
 		if (!table_status) {
@@ -92,7 +116,7 @@ const updateTableStatus = async (req, res) => {
 		res.status(200).send(response.rows)
 	} catch (err) {
 		console.log(err)
-		res.status(400).send(err)
+		res.status(500).send(err)
 	}
 }
 
@@ -101,6 +125,10 @@ const deleteTable = async (req, res) => {
 	try {
 		const { id } = req.params
 
+		if (!id) {
+			return res.status(400).send("All fields are required")
+		}
+
 		const response = await pool.query(
 			"DELETE FROM tables WHERE table_id = $1",
 			[id]
@@ -108,6 +136,7 @@ const deleteTable = async (req, res) => {
 		res.status(200).send(response.rows)
 	} catch (err) {
 		console.log(err)
+		res.status(500).send(err)
 	}
 }
